Develop

+++ PawnShopBE/PawnShopBE.Core/DTOs/BranchDTO.cs
@@ -13,7 +13,7 @@ namespace PawnShopBE.Core.DTOs
         public string BranchName { get; set; }
         public string Address { get; set; }
         public string PhoneNumber { get; set; }
-        public int Fund { get; set; }
+        public decimal Fund { get; set; }
         public int Status { get; set; }
     }
 }
